new manifest fields for permissions, bit of tidy up before pupstatus refactor
